docker-compose file to deploy congatudo
This is an example to deploy congatudo+mqtt broker using congatudo standalone installation with docker-compose. 
Keep in mind paths are relative and should be updated to your needs.
First time running it will give eacces error to congatudo.json file, if you have any, it will overwrite it with default one, you need to chmod 777 it so you can work with it,
Once done, configure it using the robot implementation, server IP for congatudo and mqtt, and user/pass for each.